third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-snapshot-comparison-expansion-preserved-when-sorting.html

<html>
<head>
  <script src="../../inspector/inspector-test.js"></script>
  <script src="../resources/heap-snapshot-test.js"></script>
<script>

function test() {
  var instanceCount = 24;
  function createHeapSnapshotA() {
    return HeapSnapshotTestRunner.createHeapSnapshot(instanceCount, 5);
  }
  function createHeapSnapshotB() {
    return HeapSnapshotTestRunner.createHeapSnapshot(instanceCount + 1, 5 + instanceCount);
  }

  HeapSnapshotTestRunner.runHeapSnapshotTestSuite([function testExpansionPreservedWhenSorting(next) {
    HeapSnapshotTestRunner.takeAndOpenSnapshot(createHeapSnapshotA, createSnapshotB);
    function createSnapshotB() {
      HeapSnapshotTestRunner.takeAndOpenSnapshot(createHeapSnapshotB, step1);
    }

    function step1() {
      HeapSnapshotTestRunner.switchToView('Comparison', step2);
    }

    function step2() {
      var row = HeapSnapshotTestRunner.findRow('B');
      TestRunner.assertEquals(true, !!row, '"B" row');
      HeapSnapshotTestRunner.expandRow(row, expandB);
      function expandB() {
        var buttonsNode = HeapSnapshotTestRunner.findButtonsNode(row);
        TestRunner.assertEquals(true, !!buttonsNode, 'no buttons node found!');
        HeapSnapshotTestRunner.clickShowMoreButton('showAll', buttonsNode, step4);
      }
    }

    var columns;
    function step4() {
      columns = HeapSnapshotTestRunner.viewColumns();
      HeapSnapshotTestRunner.clickColumn(columns[0], step5);
    }

    function step5() {
      var row = HeapSnapshotTestRunner.findRow('B');
      TestRunner.assertEquals(true, !!row, '"B" row');
      var bInstanceRow = row.children[0];
      TestRunner.assertEquals(true, !!bInstanceRow, '"B" instance row');
      HeapSnapshotTestRunner.expandRow(bInstanceRow, expandA);
      function expandA(row) {
        function propertyMatcher(node) {
          return node._referenceName === 'a' && node._name.charAt(0) === 'A';
        }
        var aRow = HeapSnapshotTestRunner.findMatchingRow(propertyMatcher, row);
        TestRunner.assertEquals(true, !!aRow, '"a: A" row');
        HeapSnapshotTestRunner.expandRow(aRow, step6);
      }
    }

    function step6() {
      var row = HeapSnapshotTestRunner.findRow('B');
      TestRunner.assertEquals(true, !!row, '"B" row');
      function deletedNodeMatcher(data) {
        return data._isDeletedNode && data._name.charAt(0) === 'B';
      }
      var bInstanceRow = HeapSnapshotTestRunner.findMatchingRow(deletedNodeMatcher, row);
      TestRunner.assertEquals(true, !!bInstanceRow, '"B" instance row');
      HeapSnapshotTestRunner.expandRow(bInstanceRow, expandA);
      function expandA(row) {
        function propertyMatcher(data) {
          return data._referenceName === 'a' && data._name.charAt(0) === 'A';
        }
        var aRow = HeapSnapshotTestRunner.findMatchingRow(propertyMatcher, row);
        TestRunner.assertEquals(true, !!aRow, '"a: A" row');
        HeapSnapshotTestRunner.expandRow(aRow, step7);
      }
    }

    var columnContents;
    function step7() {
      columnContents = HeapSnapshotTestRunner.columnContents(columns[0]);
      HeapSnapshotTestRunner.clickColumn(columns[0], clickTwice);
      function clickTwice() {
        HeapSnapshotTestRunner.clickColumn(columns[0], step8);
      }
    }

    function step8() {
      var newColumnContents = HeapSnapshotTestRunner.columnContents(columns[0]);
      HeapSnapshotTestRunner.assertColumnContentsEqual(columnContents, newColumnContents);
      setTimeout(next, 0);
    }
  }]);
}

</script>
</head>
<body onload="runTest()">
<p>
Tests Comparison view of detailed heap snapshots.
Expanded nodes must be preserved after sorting.
</p>
</body>
</html>
