third_party/WebKit/LayoutTests/http/tests/devtools/profiler/heap-snapshot-comparison-expansion-preserved-when-sorting-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests Comparison view of detailed heap snapshots. Expanded nodes must be preserved after sorting.

Profiler was enabled.
Detailed heap profiles were enabled.

Running: testExpansionPreservedWhenSorting

Profiler was disabled.
